Hoàn thành nhập thuốc, tiền và tổng tiền chuyển sang double

+++ DTO_Clinic/DTO_CTPhieuNhapThuoc.cs
@@ -9,16 +9,16 @@ namespace DTO_Clinic
     public class DTO_CTPhieuNhapThuoc:BaseModel
     {
         private int _soLuongNhap;
-        private float _donGiaNhap;
-        private float _thanhTien;
+        private double _donGiaNhap;
+        private double _thanhTien;
         private bool _isDeleted;
         public virtual DTO_PhieuNhapThuoc PhieuNhapThuoc  { get; set; }
         public virtual DTO_Thuoc Thuoc { get; set; }
         public int MaThuoc { get; set; }
         public int MaPNT { get; set; }
         public int SoLuongNhap { get => _soLuongNhap; set { _soLuongNhap = value; OnPropertyChanged(); } }
-        public float DonGiaNhap { get => _donGiaNhap; set { _donGiaNhap = value; OnPropertyChanged(); } }
-        public float ThanhTien { get => _thanhTien; set { _thanhTien = value; OnPropertyChanged(); } }
+        public double DonGiaNhap { get => _donGiaNhap; set { _donGiaNhap = value; OnPropertyChanged(); } }
+        public double ThanhTien { get => _thanhTien; set { _thanhTien = value; OnPropertyChanged(); } }
         public bool IsDeleted { get => _isDeleted; set { _isDeleted = value; OnPropertyChanged(); } }
 
         public DTO_CTPhieuNhapThuoc()
@@ -26,7 +26,7 @@ namespace DTO_Clinic
 
         }
 
-        public DTO_CTPhieuNhapThuoc(int _maPNT, int _maThuoc, int _soLuongNhap, float _donGiaNhap)
+        public DTO_CTPhieuNhapThuoc(int _maPNT, int _maThuoc, int _soLuongNhap, double _donGiaNhap)
         {
             MaPNT = _maPNT;
             MaThuoc = _maThuoc;
