+++ DTO_Clinic/DTO_Thuoc.cs
@@ -11,14 +11,14 @@ namespace DTO_Clinic
         public int Id { get; set; }
         private string _tenThuoc;
         private string _congDung;
-        private float _donGia;
+        private double _donGia;
         private int _soLuong;
         private bool _isDeleted;
         public virtual DTO_DonVi DonVi { get; set; }
         public int MaDonVi { get; set; }
         public string TenThuoc { get => _tenThuoc; set { _tenThuoc = value; OnPropertyChanged(); }  }
         public string CongDung { get => _congDung; set { _congDung = value; OnPropertyChanged(); } }
-        public float DonGia { get => _donGia; set { _donGia = value; OnPropertyChanged(); } }
+        public double DonGia { get => _donGia; set { _donGia = value; OnPropertyChanged(); } }
         public int SoLuong { get => _soLuong; set { _soLuong = value; OnPropertyChanged(); } }
         public virtual ICollection<DTO_CTPhieuKhamBenh> DS_CTPhieuKhamBenh { get; set; }
         public virtual ICollection<DTO_CTPhieuNhapThuoc> DS_CTPhieuNhapThuoc { get; set; }
@@ -30,7 +30,7 @@ namespace DTO_Clinic
             IsDeleted = false;
         }
 
-        public DTO_Thuoc(string tenThuoc, int maDonVi, float donGia, int soLuong, string congDung)
+        public DTO_Thuoc(string tenThuoc, int maDonVi, double donGia, int soLuong, string congDung)
         {
             TenThuoc = tenThuoc;
             MaDonVi = maDonVi;
